# Implement faster arrow array reader