--- a/parquet/src/arrow/record_reader.rs
+++ b/parquet/src/arrow/record_reader.rs
@@ -466,7 +466,7 @@ mod tests {
 
     impl Iterator for TestPageReader {
         type Item = Result<Page>;
-    
+
         fn next(&mut self) -> Option<Self::Item> {
             self.get_next_page().transpose()
         }