… parquet/Cargo.toml
… parquet/src/arrow/array_reader.rs
… parquet/src/arrow/arrow_array_reader.rs
… parquet/src/column/page.rs
… parquet/src/column/reader.rs
… parquet/src/encodings/mod.rs
… parquet/src/errors.rs
… parquet/src/file/serialized_reader.rs
… parquet/src/lib.rs
… parquet/src/schema/types.rs
… parquet/src/util/memory.rs
… parquet/src/util/mod.rs
… parquet/src/util/test_common/page_util.rs