@nevi-me 2021-06-04
> @nevi-me is this something you can take on reviewing / approving? I am not very familiar with this code -- it looked good to me but I don't feel super confident of approving it. However, if you don't have the time I will do the best I can

I'll complete my review over the weekend, I like the approach; the `RefCell` makes it feel complicated, but I appreciate why.

I don't think I'll have any major items to raise though