@alamb 2021-06-08
FYI I plan to make a release candidate for Arrow 4.3 on Thursday or Friday this week and release early next week. So if we want to try and get this PR into 4.3 that is the schedule.

It is large enough, however, that delaying until 4.4 and giving it some more bake time is not a bad idea either