@alamb 2021-06-10
Merging this in and we can figure out if we want to try and put it into arrow 4.4.0. Thanks @yordan-pavlov !